requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.1.0
dateparser==1.1.8
emoji==2.2.0
exceptiongroup==1.1.1
flake8==6.0.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.4
iniconfig==2.0.0
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.9
pybit==5.2.0
pycodestyle==2.10.0
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.0.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
tomli==2.0.1
types-requests==2.31.0.20240125
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==2.1.0
websocket-client==1.5.1
websockets==11.0.2
yarl==1.9.2
